nm: Move profile activation to the connection module
The profile activation handlers have been relocated to the
connection.ConnectionProfile class.

The ability to activate a device from the device module is kept to allow
callers to migrate in following patches.
