+++ libnmstate/nm/device.py
@@ -21,215 +21,10 @@ from . import connection
 from . import nmclient
 
 
-class AlternativeACState(object):
-    UNKNOWN = 0
-    ACTIVE = 1
-    ACTIVATING = 2
-    FAIL = 3
-
-
-class ActiveConnection(object):
-
-    def __init__(self, active_connection):
-        self._act_con = active_connection
-        nmdevs = active_connection.get_devices()
-        self._nmdev = nmdevs[0] if nmdevs else None
-
-        self.handlers = set()
-
-        self.refresh_state()
-
-    def refresh_state(self):
-        self._state = self._act_con.get_state()
-        self._state_reason = self._act_con.get_state_reason()
-        self._alternative_state = AlternativeACState.UNKNOWN
-
-        nm_acs = nmclient.NM.ActiveConnectionState
-        nm_acsreason = nmclient.NM.ActiveConnectionStateReason
-        if self._state == nm_acs.DEACTIVATED:
-            unable_to_activate = (
-                    not self._nmdev or
-                    (
-                        self._state_reason is not None and
-                        self._state_reason != nm_acsreason.DEVICE_DISCONNECTED
-                    ) or
-                    self._nmdev.get_active_connection() is not self._act_con
-            )
-            if unable_to_activate:
-                self._alternative_state = AlternativeACState.FAIL
-            # Use the device-state as an alternative to determine if active.
-            elif (self.nmdev_state <= nmclient.NM.DeviceState.DISCONNECTED or
-                    self.nmdev_state > nmclient.NM.DeviceState.DEACTIVATING):
-                self._alternative_state = AlternativeACState.FAIL
-
-    @property
-    def is_active(self):
-        nm_acs = nmclient.NM.ActiveConnectionState
-        if self._state == nm_acs.ACTIVATED:
-            return True
-        elif self._state == nm_acs.ACTIVATING:
-                # master connections qualify as activated once they
-                # reach IP-Config state. That is because they may
-                # wait for slave devices to attach
-                return (
-                    _is_device_master_type(self._nmdev) and
-                    nmclient.NM.DeviceState.IP_CONFIG <= self.nmdev_state <=
-                    nmclient.NM.DeviceState.ACTIVATED
-                )
-
-        return False
-
-    @property
-    def is_activating(self):
-        activation_failed = self._alternative_state == AlternativeACState.FAIL
-        return not self.is_active and not activation_failed
-
-    @property
-    def reason(self):
-        return self._state_reason
-
-    @property
-    def nm_active_connection(self):
-        return self._act_con
-
-    @property
-    def devname(self):
-        return self._nmdev.get_iface()
-
-    @property
-    def state(self):
-        return self._state
-
-    @property
-    def nmdev_state(self):
-        return (self._nmdev.get_state() if self._nmdev
-                else nmclient.NM.DeviceState.UNKNOWN)
-
-
 def activate(dev=None, connection_id=None):
     """Activate the given device or remote connection profile."""
-    mainloop = nmclient.mainloop()
-    mainloop.push_action(
-        _safe_activate_async, dev, connection_id)
-
-
-def _safe_activate_async(dev, connection_id):
-    client = nmclient.client()
-    mainloop = nmclient.mainloop()
-    cancellable = mainloop.new_cancellable()
-
     conn = connection.ConnectionProfile()
-    conn.import_by_id(connection_id)
-    if conn.profile:
-        dev = None
-    active_conn = connection.get_device_active_connection(dev)
-    if active_conn:
-        ac = ActiveConnection(active_conn)
-        if ac.is_activating:
-            logging.debug(
-                'Connection activation in progress: dev=%s, state=%s',
-                ac.devname, ac.state)
-            _waitfor_active_connection_async(ac, mainloop)
-            return
-
-    specific_object = None
-    user_data = mainloop, dev, connection_id, cancellable
-    client.activate_connection_async(
-        conn.profile,
-        dev,
-        specific_object,
-        cancellable,
-        _active_connection_callback,
-        user_data,
-    )
-
-
-def _active_connection_callback(src_object, result, user_data):
-    mainloop, nmdev, connection_id, cancellable = user_data
-    mainloop.drop_cancellable(cancellable)
-
-    try:
-        nm_act_con = src_object.activate_connection_finish(result)
-    except Exception as e:
-        activation_type, activation_object = _get_activation_metadata(
-            nmdev, connection_id)
-
-        if mainloop.is_action_canceled(e):
-            logging.debug(
-                'Connection activation canceled on %s %s: error=%s',
-                activation_type, activation_object, e)
-        elif is_connection_unavailable(e):
-            logging.warning('Connection unavailable on %s %s, retrying',
-                            activation_type, activation_object)
-            mainloop.execute_last_action()
-        else:
-            mainloop.quit(
-                'Connection activation failed on {} {}: error={}'.format(
-                    activation_type, activation_object, e))
-        return
-
-    if nm_act_con is None:
-        activation_type, activation_object = _get_activation_metadata(
-            nmdev, connection_id)
-        mainloop.quit('Connection activation failed on %s %s: error=unknown' %
-                      activation_type, activation_object)
-    else:
-        devname = nm_act_con.props.connection.get_interface_name()
-        logging.debug('Connection activation initiated: dev=%s, con-state=%s',
-                      devname, nm_act_con.props.state)
-
-        ac = ActiveConnection(nm_act_con)
-        if ac.is_active:
-            mainloop.execute_next_action()
-        elif ac.is_activating:
-            _waitfor_active_connection_async(ac, mainloop)
-        else:
-            mainloop.quit(
-                'Connection activation failed on {}: reason={}'.format(
-                    ac.devname, ac.reason))
-
-
-def is_connection_unavailable(err):
-    return (isinstance(err, nmclient.GLib.GError) and
-            err.domain == 'nm-manager-error-quark' and
-            err.code == 2 and
-            'is not available on the device' in err.message)
-
-
-def _get_activation_metadata(nmdev, connection_id):
-    if nmdev:
-        activation_type = 'device'
-        activation_object = nmdev.get_iface()
-    elif connection_id:
-        activation_type = 'connection_id'
-        activation_object = connection_id
-    else:
-        activation_type = activation_object = 'unknown'
-
-    return activation_type, activation_object
-
-
-def _waitfor_active_connection_async(ac, mainloop):
-    ac.handlers.add(
-        ac.nm_active_connection.connect(
-            'state-changed', _waitfor_active_connection_callback, ac, mainloop)
-    )
-
-
-def _waitfor_active_connection_callback(
-        nm_act_con, state, reason, ac, mainloop):
-    ac.refresh_state()
-    if ac.is_active:
-        logging.debug('Connection activation succeeded: dev=%s, con-state=%s',
-                      ac.devname, ac.state)
-        for handler_id in ac.handlers:
-            ac.nm_active_connection.handler_disconnect(handler_id)
-        mainloop.execute_next_action()
-    elif not ac.is_activating:
-        for handler_id in ac.handlers:
-            ac.nm_active_connection.handler_disconnect(handler_id)
-        mainloop.quit('Connection activation failed on {}: reason={}'.format(
-            ac.devname, ac.reason))
+    conn.activate(dev, connection_id)
 
 
 def deactivate(dev):
@@ -423,16 +218,3 @@ def get_device_common_info(dev):
         'type_name': dev.get_type_description(),
         'state': dev.get_state(),
     }
-
-
-def _is_device_master_type(nmdev):
-    if nmdev:
-        gobject = nmclient.GObject
-        is_master_type = (
-            gobject.type_is_a(nmdev, nmclient.NM.DeviceBond) or
-            gobject.type_is_a(nmdev, nmclient.NM.DeviceBridge) or
-            gobject.type_is_a(nmdev, nmclient.NM.DeviceTeam) or
-            gobject.type_is_a(nmdev, nmclient.NM.DeviceOvsBridge)
-        )
-        return is_master_type
-    return False
